Added functionality to AbstractDatabaseManager to allow it to create and terminate connections to the MySQL database and to create the database upon the first attempt to connect to it.
Currently, I have everything set up as follows:
	server host: 	 localhost
	database name:	 facultypurchaseorders
	user name: 		 root (signifies admin priviledge)
	password:		 cap-4710-project-spring-2014
If you want to install MySQL and run the main method in
AbstractDatabaseManager, you might want to use these settings until we
know the actual server we'll be using (at which point I'll change the
code). Don't worry about creating the database. As long as the program
can connect to the host, it will create the database the first time it
connects.

I also renamed SystemLogger to LoggerManager and removed unnecessary
logger functionality from it.