Added printing and inserting functionality to AbstractDatabaseManager Added functionality to AbstractDatabaseManager for printing the contents of the USERS table to the file "output/cdContents.csv" (can be opened in WordPad, MS Excel, or LibreOffice Calc) and inserting table entries to their corresponding tables in the database.
Created a new abstract class named AbstractTableEntry used to represent
entries in the database tables and maintain SQL data surrounding them
(such as column names).

Refactored Users class to be named "User" and modified the class to act
as an extension of AbstractTableEntry so it could be used in
communications with the database.