Added Tomcat Third Tarty Files and Updated Build Path
I've added an installation of Tomcat to the 'external' folder in the
folder 'apache-tomcat-7.0.52'. I've also included a file named
ECLIPSE-CONFIGURATION.txt within that folder that gives steps on how to
configure Eclipse to use Tomcat and how to run Tomcat from Eclipse. If
you still see errors with your files after this, Jonathan, let me know.
The project now has access to all of Tomcat's java classes, but that
doesn't necessarily mean everything will start working. 

I also updated the database logic to allow entries in the ORDERS table
to be treated as AbstractTableEntry objects, which means we can now
insert, delete, update, and select orders from the database.