Added Functionality for Finding Database Entries by Column Values
I added a selectRowsByColumns method to AbstractDatabaseManager that
takes in a HashMap of column names to their respective values and a
table entry class and returns a List of AbstractTableEntries
representing all the database entries that match those column values.
Basically, this will allow us to look for things like usernames and
passwords in the database tables and return all the entries that contain
them.

I also modified Order, User, and AbstractDatabaseManager to fix some
MySQL data type issues I was running into.

(If Git doesn't show all the actual changes, its because I somehow got
disconnected from the 'master' branch. On a side note, if Git ever tells
you something along the lines of "current branch is not configured for
pull" when you try to pull or fetch from the repository, try the
following:

1. Open the cap-4710-project\.git folder. It might be hidden, so you
might want to make hidden files visible first.

2. Open the 'config' file and add the following after the last line:

[branch "master"]
    remote = origin
    merge = refs/heads/master
    
This will tell Git what branch to use for merging and should fix the
problem.)