Added Method for Creating and Verifying Purchase Orders
Modified ServerImpl to add a method for creating and verifying purchase
orders from the information put into the fields of
FacultyOrderForm.html. This method isn't quite done yet, but at the very
least, it verifies to make sure the requestor is in the database, adds
an order request with the passed in information to the database, and
sends an email to the requestor verifying that their order has been
sent.

On a side note, I still need to figure out how I'm going to handle files
in the database and get newly created order ID's, so I'll be working on
adding functionality for both of these tomorrow.